test:Calculate expected result
test test_fake_transactions_returns_same_data_with_same_seed
is currently failing in CI
https://github.com/jamiekt/jstark/actions/runs/8038163407/job/21953772143#step:7:226

Strangely it succeeds on Windows but fails on Linux and MacOS. I have a
theory that Faker is behaving different on those platforms and is
returning different fake transactions.
For that reason, calculate exactly what the result should be from the
fake transactions and ensure the value calculated by kstark matches.